requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
flask-swagger-ui==4.11.1
APIFlask==1.3.1


Jinja2==3.1.2
marshmallow==3.19.0
psycopg2-binary==2.9.6


bcrypt==4.0.1
cryptography==41.0.1
PyJWT==2.8.0


celery==5.3.1
kombu==5.3.1


python-dotenv==1.0.0
gunicorn==20.1.0

marshmallow-sqlalchemy==0.29.0
jsonschema==4.19.1